refactor: merge import
